Consolidate unit tests for fonduer.utils.visualizer and rename the file to tests/utils/test_visualizer.py
